refactor: Car 관련 비즈니스 로직 Car 클래스에 캡슐화
