+++ src/main/java/racingcar/domain/Car.java
@@ -1,6 +1,9 @@
 package racingcar.domain;
 
-import static racingcar.constant.CarConstant.MAX_CAR_NAME_LENGTH;
+import camp.nextstep.edu.missionutils.Randoms;
+
+import static racingcar.constant.CarConstant.*;
+import static racingcar.constant.CarConstant.CAR_MOVE_FORWARD_STANDARD;
 import static racingcar.view.message.ErrorMessage.CAR_NAME_LENGTH_ERROR_MESSAGE;
 
 public class Car implements Comparable<Car> {
@@ -18,8 +21,8 @@ public class Car implements Comparable<Car> {
     }
 
     // 추가 기능 구현
-    public void move(int position) {
-        this.position += position;
+    public void move() {
+        this.position += getMovingPosition();
     }
 
     public String getName() {
@@ -34,6 +37,17 @@ public class Car implements Comparable<Car> {
         if (name.length() > MAX_CAR_NAME_LENGTH) {
             throw new IllegalArgumentException(CAR_NAME_LENGTH_ERROR_MESSAGE);
         }
+    }
+
+    private Integer getMovingPosition() {
+        int randomMovingValue = Randoms.pickNumberInRange(MIN_CAR_MOVE_FORWARD_RANDOM_VALUE, MAX_CAR_MOVE_FORWARD_RANDOM_VALUE);
+        Integer movingPosition = 0;
+
+        if (randomMovingValue >= CAR_MOVE_FORWARD_STANDARD) {
+            return 1;
+        }
+
+        return movingPosition;
     }
 
     @Override
